+++ SolidModelBrowser/App.xaml.cs
@@ -1,18 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
 using System.Windows;
 
 namespace SolidModelBrowser
 {
-    /// <summary>
-    /// Логика взаимодействия для App.xaml
-    /// </summary>
     public partial class App : Application
     {
-
         public static void SetTheme(string theme)
         {
             Current.Resources.Clear();